% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/between_emd_fun.R
\name{between_emd}
\alias{between_emd}
\title{Calculates the between seasonal range Earth-mover's distance (EMD)}
\usage{
between_emd(t1, t2)
}
\arguments{
\item{t1}{A list of rasters for each interval for a given range.}

\item{t2}{A second list of raster for each interval for a given range that
is being compared with the first list of rasters (t1).}
}
\value{
The output is a list of matrices that contain the EMDs for each
individual in a given year comparing different intervals.
}
\description{
Calculates the between seasonal range Earth-mover's distance (EMD)
}
