% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_intervals_fun.R
\name{check}
\alias{check}
\title{Grabs the attributes that you assigned to a list}
\usage{
check(list, attribute_name)
}
\arguments{
\item{list}{The list object that the attribute is being assigned to.}

\item{attribute_name}{The name of the attribute item created for both
\code{list1} and \code{list2} The input for this argument must be a character.}
}
\value{
This grabs the attributes that you assigned to the list.
}
\description{
Grabs the attributes that you assigned to a list
}
