% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emd_env_fun.R
\name{emd_env}
\alias{emd_env}
\title{Process UD raster to calculate EMD with pacakge 'transport'}
\usage{
emd_env(rs1, rs2, nbins = 50)
}
\arguments{
\item{rs1}{A RasterStack with two layers: (1) the rasterized UD and (2) the
values of the environmental covariate.}

\item{rs2}{A RasterStack with two layers: (1) the rasterized UD and (2) the
values of the environmental covariate.}

\item{nbins}{a numeric vector of length == 1 giving the number of bins to
discretize environmental space into. Will create a sequence from
the min -- max environmental values (from both ranges combined).
Defaults to 50.}
}
\value{
EMD between rs1 and rs2 in one dimension of environmental space.
}
\description{
Process UD raster to calculate EMD with pacakge 'transport'
}
