% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emd_geo_fun.R
\name{emd_geo}
\alias{emd_geo}
\title{Process UD raster to calculate EMD}
\usage{
emd_geo(r1, r2)
}
\arguments{
\item{r1}{is a RasterLayer representing a rasterized UD.}

\item{r2}{is a RasterLayer representing a rasterized UD.}
}
\value{
Function calculates EMD between r1 and r2 in geographic space.
}
\description{
Process UD raster to calculate EMD
}
