% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMigD.R
\docType{package}
\name{EMigD}
\alias{EMigD}
\title{EMigD: A package for classifying migrants}
\description{
This package provides a means to quantify an individual's
migratory status within a given year. This packages allows the user
to determine an individual's migratory status within geographical
and environmental space.
}
