% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMigD_format_fun.R
\name{EmigD_format}
\alias{EmigD_format}
\title{Re-formats the User's data frame for the EMigD workflow}
\usage{
EmigD_format(dat, id_, x_, y_, year_, month_, jdate_)
}
\arguments{
\item{dat}{A data frame}

\item{id_}{The individual ID column in 'dat'. This column should be a 'factor'.}

\item{x_}{The x-coordinate column in 'dat'.}

\item{y_}{The y-coordinate column in 'dat'.}

\item{year_}{The column containing the years in 'dat'. The column should be
a numeric value.}

\item{month_}{The column containing the months in 'dat'. The column should
be a numeric value.}

\item{jdate_}{The column containing the julian dates. This column should be
a numeric value}
}
\value{
Function re-formats the user's data frame for the EmigD workflow
}
\description{
Re-formats the User's data frame for the EMigD workflow
}
