% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emspeeds_fun.R
\name{emspeeds}
\alias{emspeeds}
\title{Function arguments #}
\usage{
emspeeds(t1, t2, emd)
}
\arguments{
\item{t1}{First list of data frames, where each element represents an interval
or time frame (e.g., first 10-day period to second 10-day period or
winter to summer)}

\item{t2}{Second list of data frames, where each element represents an interval or
time frame (e.g., first 10-day period to second 10-day period or winter to
summer)

@returns Matrices of the emspeeds (EMDs/Difference between the two)}

\item{emd}{Either the between time frame (e.g., winter season to the summer season)
emd or the within-time frame (e.g., first 10-days to the second 10-days) emd}
}
\description{
Function arguments #
}
