% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/within_emd_fun.R
\name{within_emd}
\alias{within_emd}
\title{Calculates the Earth-Mover's Distance (EMD) within the same range}
\usage{
within_emd(t)
}
\arguments{
\item{t}{List of rasters containing the raster for each period within the}

\item{nintervals}{The length that the rasters are broken up into by the
specified interval.}
}
\value{
The output is a list of matrices that contain the within- range EMDs
}
\description{
Calculates the Earth-Mover's Distance (EMD) within the same range
}
