% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_intervals_fun.R
\name{match_intervals}
\alias{match_intervals}
\title{Matches the number of intervals between the two seasons selected}
\usage{
match_intervals(list1, list2)
}
\arguments{
\item{list1}{A list object that is the reference for \code{list2}}

\item{list2}{A list object that the length of \code{list1} is being matched to}
}
\value{
The output is a list that has the same intervals as that in the
\code{list2} object.
}
\description{
Matches the number of intervals between the two seasons selected
}
